+++ modules/core/src/main/java/org/vstu/compprehension/models/repository/QuestionMetadataBaseRepository.java
@@ -17,15 +17,15 @@ public interface QuestionMetadataBaseRepository <T extends QuestionMetadataEntit
     @Query("select q from #{#entityName} q where q.stage = 3")  // Note: db field `_stage` mapped by entity to `stage`
     List<QuestionMetadataEntity> findAllReady();
 
-//    @Query("select q from #{#entityName} q where q.stage = 3 and q.concept_bits IN :conceptBitEntries")  // Note: db field `_stage` mapped by entity to `stage`
-//    List<QuestionMetadataBaseEntity> findAllWithConcepts(Collection<Integer> conceptBitEntries, Pageable pageable);
+//    @Query("select q from #{#entityName} q where q.stage = 3 and q.concept_bits Long:conceptBitEntries")  // Note: db field `_stage` mapped by entity to `stage`
+//    List<QuestionMetadataBaseEntity> findAllWithConcepts(Collection<Long> conceptBitEntries, Pageable pageable);
 
     @Query("select q from #{#entityName} q where q.stage = 3 AND q.conceptBits IN :values")  // Note: db field `concept_bits` mapped by entity to `conceptBits`
-    List<QuestionMetadataEntity> findAllWithConcepts(@Param("values") Collection<Integer> conceptBitEntries);
+    List<QuestionMetadataEntity> findAllWithConcepts(@Param("values") Collection<Long> conceptBitEntries);
 
     @Query("select q from #{#entityName} q where q.stage = 3 AND q.conceptBits IN :values AND q.templateId NOT IN :ids")
     List<QuestionMetadataEntity> findAllWithConceptsWithoutTemplates(
-            @Param("values") Collection<Integer> conceptBitEntries,
+            @Param("values") Collection<Long> conceptBitEntries,
             @Param("ids") Collection<Integer> templatesIds
     );
